implement custom permission system

+++ src/commands/stickers/deletesticker.ts
@@ -1,8 +1,24 @@
-import { SlashCommandBuilder } from "discord.js";
+import {
+  ApplicationIntegrationType,
+  InteractionContextType,
+  MessageFlags,
+  SlashCommandBuilder,
+} from "discord.js";
 import type { CommandExecutor } from "../../types/commands.js";
 import { deleteSticker } from "../../db/index.js";
+import { isUserAllowed } from "../../utils/users.js";
+import { Constants } from "../../utils/index.js";
 
 export const data = new SlashCommandBuilder()
+  .setContexts([
+    InteractionContextType.PrivateChannel,
+    InteractionContextType.Guild,
+    InteractionContextType.BotDM,
+  ])
+  .setIntegrationTypes([
+    ApplicationIntegrationType.GuildInstall,
+    ApplicationIntegrationType.UserInstall,
+  ])
   .setName("deletesticker")
   .setDescription("Delete a sticker by title")
   .addStringOption((opt) =>
@@ -14,6 +30,13 @@ export const data = new SlashCommandBuilder()
   );
 
 export const execute: CommandExecutor = async (interaction) => {
+  if (!(await isUserAllowed("deleteSticker", interaction))) {
+    return interaction.reply({
+      content: Constants.PERMISSION_PUNT_MESSAGE,
+      flags: MessageFlags.Ephemeral,
+    });
+  }
+
   const id = interaction.options.getString("query", true);
   try {
     const isDeleted = await deleteSticker(id);
